little fix Product Hub gRPC test

See merge request qa/automag!915

+++ web-tests/grpc/src/test/java/ru/instamart/test/content/product_hub/ProductHubBackTest.java
@@ -448,7 +448,7 @@ public class ProductHubBackTest extends GrpcBase {
                         .addOfferData(Data.newBuilder()
                                 .setKey("popularity")
                                 .addValues("20")
-                                .setType(ValueType.DOUBLE)
+                                .setType(ValueType.INTEGER)
                                 .build())
                         .setStatus(Status.ENABLE)
                         .build())
